Add subcell lighting; separate walk-through from see-through
Separating walk-through from see-through is important because the player cannot be walked through, but he needs to be able to see from the center of his tile out through two layers of subcells into the surrounding environment. Cells that cannot be walked through are given a 5x5 block of solid subcells.